### 2016-08-25 #54 @whedon
Submission: cartography: Create and Integrate Maps in your R Workflow
**Submitting author:** @rCarto (<a href="http://orcid.org/0000-0002-1932-3323">Timothée Giraud</a>)
**Repository:** <a href="https://github.com/Groupe-ElementR/cartography" target ="_blank">https://github.com/Groupe-ElementR/cartography</a>
**Version:** v1.4.0
