### 2016-08-25 #54 @janschulz
* Downloaded from the repo via devtools :heavy_check_mark:
* License is GPL (form 2007 -> DESCRIPTION says GPL-3) :heavy_check_mark:
* redid the example in the readme in an jupyter notebook on R 3.3 :heavy_check_mark: The vignette also worked like a charm :heavy_check_mark:  Unfortunatelly the only EN blogpost in http://rgeomatic.hypotheses.org/842 resulted in some errors (deprecation warnings -> the blogpost seems to require 1.3 but I was on 1.4)
* no performance clames were given, but the examples were reasonable (compared to cartopy/basemaps in python)
* the paper states what the packages tries to solve  :heavy_check_mark:
* the (devtools) installation worked flawless on windows so dependencies are automatically installed :heavy_check_mark:
* the examples solve real world needs  :heavy_check_mark:
* checked the API docs and they look great  :heavy_check_mark:
* The package is auto checked on travis (Linux)  :heavy_check_mark: (as a windows user I would appreciate appveyor tests as well :-) )
* community guidelines are basically "contribute vie PRs, raise issues or ask questions via issues" -> Minimum is answered  :heavy_check_mark:
* [paper](https://github.com/Groupe-ElementR/cartography/tree/master/paper) contains a list of authors with affiliations (and one ORCID :+1: ) :heavy_check_mark:
* the references only cite packages or books and an R NEws item, which seems to have no DOI :-( :heavy_check_mark:

So IMO there is only one question:

* Version on github is 1.4.0 but CRAN is still on 1.3.0 -> whats the problem here? https://github.com/Groupe-ElementR/cartography/releases vs https://cran.r-project.org/web/packages/cartography/index.html

Apart form that this looks good to publish!
