### 2016-08-26 #54 @rCarto
Thanks for the review @janschulz.
* I have updated the blogpost to reflect the current version usage.
* I have added the appveyor CI (which was surprisingly easy).

@arfon, concerning the version numbers, I was waiting for possible comments or suggestions from JOSS to integrate. Since comments and suggestions do not imply code changes I will upload this release (v1.4.0) to the CRAN today.
