### 2016-08-26 #54 @janschulz
> I have updated the blogpost to reflect the current version usage.
> I have added the appveyor CI (which was surprisingly easy).

Thank you! :tada:

[Just to make sure: both were just something I observed and not related to the review here. Both could have gone as easily to the upstream repo as a issue -- and probably should have... sorry about that]
