### 2016-08-27 #54 @rCarto
cartography v1.4.0 is on CRAN now.
