### 2016-08-27 #54 @arfon
@rcarto - could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
