### 2016-08-30 #54 @arfon
Thanks for the rapid review @janschulz :zap:

@rCarto your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00054 🚀 🎉 💥
